Issue#35 fix removed favourite bug

+++ client/src/pages/components/EnergizerDeleteButton.js
@@ -1,27 +1,48 @@
 import React, { useState } from "react";
 
-function EnergizerDeleteButton({ energizerId }) {
+function EnergizerDeleteButton({ energizerId, energizerName }) {
 	const [isDeleting, setIsDeleting] = useState(false);
+	// this function validates whether the energizer is in the local storage
+	function findFavouriteEnergizer() {
+		let favourites_array = localStorage.getItem("favourite");
+		favourites_array = JSON.parse(favourites_array);
+		return favourites_array.find((i) => i.id === +energizerId);
+	}
 
 	const handleDelete = async (e) => {
 		e.preventDefault();
-		alert("Do you want to delete this energizer?");
-		try {
-			setIsDeleting(true);
-			const response = await fetch(`/api/energizers/${energizerId}`, {
-				method: "DELETE",
-			});
-			if (!response.ok) {
-				const errorText = await response.text();
-				throw new Error(errorText);
-			}
+		// if the energizer is in the favourites/local storage alerts the user and asks to remove it from favourites
+		if (findFavouriteEnergizer()) {
+			alert(
+				"Please remove the Energizer form your favourites before you delete this energizer"
+			);
+			return;
+		}
+		//confirmation box to validate whether you really want to remove the energizer, prevents accidental deletes
+		if (
+			window.confirm(
+				`Do you really want to delete energizer: ${energizerName}?`
+			)
+		) {
+			//fixed the alert, the alert now displays the name of the energizer to be removed
+			alert(`You are deleting energizer: ${energizerName}`);
+			try {
+				setIsDeleting(true);
+				const response = await fetch(`/api/energizers/${energizerId}`, {
+					method: "DELETE",
+				});
+				if (!response.ok) {
+					const errorText = await response.text();
+					throw new Error(errorText);
+				}
 
-			window.history.back();
-		} catch (error) {
-			console.error(error);
-			alert("Failed to delete energizer");
-		} finally {
-			setIsDeleting(false);
+				window.history.back();
+			} catch (error) {
+				console.error(error);
+				alert("Failed to delete energizer");
+			} finally {
+				setIsDeleting(false);
+			}
 		}
 	};
